Fix normalization of single entity responses.
A bug in the normalization logic was causing models to show up multiple
times, one that didn't have all of the attributes properly set. As it
turns out, Ember.js is trying to be smart and keeping a local cache of
records that it knows about. If the POST response isn't parsed
correctly, this local cache becomes corrupt, and appears to show
multiple copies of a record.